// src/app/proto/s2c.d.ts
export namespace S2C {

    interface IMessage {
// ...
        SC_ROLE_OFFLINE?: (S2C.ISC_ROLE_OFFLINE|null);
    }
// ...
    interface ILOGIN_SC_CHOOSE_SERVER {
        ip?: (string|null);
        port?: (number|null);
        version?: (string|null);
        resVersion?: (string|null);
        resServerAddress?: (string|null);
        uid?: (number|null);
        token?: (string|null);
        serverId?: (number|null);
        serverName?: (string|null);
    }
// ...
    interface ISC_ROLE_OFFLINE {
        type?: (number|null);
    }
// ...
}